requirements.txt
Django>=3.1.2
djangorestframework>=3.12.1,<3.13.0
djangorestframework-simplejwt>=4.6.0,<4.7.0
psycopg2>=2.8.6,<2.8.7